Profiling now uses our_math_lib
